+++ package.json
@@ -41,6 +41,9 @@
     "raw-loader": "^0.5.1",
     "react": "^0.13.3",
     "react-hot-loader": "^1.3.0",
+    "react-redux": "^2.1.2",
+    "redux": "^3.0.0",
+    "redux-devtools": "^2.1.2",
     "timeago": "^0.2.0",
     "webpack": "^1.12.1",
     "webpack-dev-server": "^1.10.1"
