+++ resources/views/files.twig
@@ -1,8 +1,5 @@
 <style>
 
-    {# Common plugin CSS. #}
-    {% include common_css_path %}
-
 
     {#
         Configurable CSS values provided through WP settings panel.
@@ -22,12 +19,21 @@
 </style>
 
 {% if environment == "development" %}
+
   <h1 style="text-align: center; color:red; font-weight:bold;">DEVELOPMENT MODE</h1>
   <script src="http://localhost:8082/webpack-dev-server.js"></script>
   <script src="http://localhost:8082/static/app.js"></script>
+
 {% else %}
+
   <script>
     {{js_string|raw}}
   </script>
+
+  <style>
+    {# Common plugin CSS. #}
+    {% include common_css_path %}
+  </style>
+
 {% endif %}
