+++ webpack.config.js
@@ -13,7 +13,7 @@ module.exports = {
   },
   module: {
     loaders: [
-      { test: /\.styl$/, loader: ExtractTextPlugin.extract("style", "css!stylus")},
+      { test: /\.styl$/, loader: 'style-loader!css-loader!stylus-loader?paths[]=./styles&paths[]=./node_modules'},
       { test: /\.svg$/, exclude:'node_modules', loader: 'raw-loader' },
       { test: /\.json$/, loader: 'json-loader' },
       { test: /\.js$/, exclude:'node_modules', loader: 'babel-loader?stage=0&loose=all' }
