+++ webpack.development.config.js
@@ -21,7 +21,7 @@ module.exports = {
   },
   module: {
     loaders: [
-      { test: /\.styl$/, loader: ExtractTextPlugin.extract("style", "css!stylus")},
+      { test: /\.styl$/, loader: 'style-loader!css-loader!stylus-loader?paths[]=./src/styles&paths[]=./node_modules'},
       { test: /\.svg$/, exclude:'node_modules', loader: 'raw-loader' },
       { test: /\.json$/, loader: 'json-loader' },
       { test: /src\/.*\.jsx?$/, loaders: ['react-hot', 'babel-loader?stage=0&loose=all'], exclude: '/node_modules/',  include: path.join(__dirname, "src") }
